requirements.txt
blinker==1.6.2
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
fonttools==4.42.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
nltk==3.8.1
numpy==1.26.0
packaging==23.1
pandas==2.1.0
Pillow==10.0.1
psycopg2-binary==2.9.7
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.8.8
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
SQLAlchemy==2.0.21
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==2.3.7
